app/requirements.txt
Flask>=3.0.0
pytesseract>=0.3.10
Pillow>=10.4.0
pdf2image>=1.17.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
ollama>=0.1.0
